Update to match v1.9.0 labymod 3 release

+++ core/src/main/java/com/rettichlp/unicacityaddon/base/utils/TextUtils.java
@@ -35,7 +35,7 @@ public class TextUtils {
         // TODO: 19.12.2022 Laby.labyAPI().renderPipeline().componentRenderer().plainSerializer() ?
     }
 
-    public static String parseTimer(int seconds) {
+    public static String parseTimer(long seconds) {
         return seconds >= 3600 ? String.format("%02d:%02d:%02d", seconds / 3600, seconds % 3600 / 60, seconds % 60) : String.format("%02d:%02d", seconds / 60, seconds % 60);
     }
